requirements.txt
torch>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
tqdm>=4.65.0
pyyaml>=6.0
# transformers는 사전학습 모델 사용 금지로 인해 제거됨
